+++ qiskit/ignis/experiments/calibration/__init__.py
@@ -12,5 +12,4 @@
 
 """Qiskit Ignis calibration module."""
 
-from .cal_metadata import CalibrationMetadata
 from .cal_base_analysis import Calibration1DAnalysis
